+++ src/ilPSP/layer_1.2-ilPSP/ilPSP.Connectors.Matlab/BatchmodeConnector.cs
@@ -559,6 +559,8 @@ namespace ilPSP.Connectors.Matlab {
         ProcessStartInfo psi;
 
 
+        bool SuccessfulExe = false;
+
         /// <summary>
         /// executes all commands
         /// </summary>
@@ -566,6 +568,7 @@ namespace ilPSP.Connectors.Matlab {
         /// a reader to the standard output of the MATLAB process.
         /// </returns>
         public void Execute(bool PrintOutput = true) {
+            SuccessfulExe = false;
             ilPSP.MPICollectiveWatchDog.Watch(csMPI.Raw._COMM.WORLD);
             if (Executed == true)
                 throw new InvalidOperationException("Execute can be called only once.");
@@ -670,6 +673,8 @@ namespace ilPSP.Connectors.Matlab {
                     throw new NotImplementedException("output object type not implemented.");
                 }
             }
+
+            SuccessfulExe = true;
         }
 
 
@@ -695,11 +700,15 @@ namespace ilPSP.Connectors.Matlab {
         /// Kills the MATLAB process
         /// </summary>
         public void Dispose() {
-            foreach (var f in CreatedFiles) {
-                File.Delete(f);
-            }
-            if (DelWorkingDir) {
-                Directory.Delete(WorkingDirectory.FullName, true);
+            if (SuccessfulExe) {
+                foreach (var f in CreatedFiles) {
+                    File.Delete(f);
+                }
+                if (DelWorkingDir) {
+                    Directory.Delete(WorkingDirectory.FullName, true);
+                }
+            } else {
+                // keeping files for diagnostic purposes
             }
         }
     }
